requirements.txt
google-generativeai
flask
gunicorn
python-dotenv